moved some things, filter in drawer

+++ src/app/modules/keybinds/keybinds.component.ts
@@ -2,23 +2,29 @@
 import { Component, ViewEncapsulation, OnInit } from '@angular/core';
 import { RouterLink } from '@angular/router';
 
-import { MatInputModule } from '@angular/material/input';
-import { MatFormFieldModule } from '@angular/material/form-field';
 import { ReactiveFormsModule, FormsModule, FormBuilder, FormGroup, Validators } from '@angular/forms';
+
+import { Subject, takeUntil } from 'rxjs';
 
 //Angular Material
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatDialogModule, MatDialog } from '@angular/material/dialog';
 
 //Components
 import { KeyboardComponent } from './keyboard/keyboard.component';
 import { AbilitiesComponent } from './abilities/abilities.component';
 import { KeybindsDrawerComponent } from './keybinds-drawer/keybinds-drawer.component';
+import { ConfirmDialogComponent } from 'app/core/components/confirm-dialog.component';
 
 //Services
 import { KeybindingService } from 'app/core/services/keybinding.service';
+import { AuthService } from 'app/core/auth/auth.service';
 @Component({
     selector: 'keybinds',
     templateUrl: './keybinds.component.html',
@@ -36,13 +42,16 @@ import { KeybindingService } from 'app/core/services/keybinding.service';
         MatSidenavModule,
         MatFormFieldModule,
         MatInputModule,
+        MatTooltipModule,
 
         KeyboardComponent,
         AbilitiesComponent,
-        KeybindsDrawerComponent
+        KeybindsDrawerComponent,
     ],
 })
 export class KeybindsComponent implements OnInit {
+    isAuthenticated: boolean;
+
     nameForm: FormGroup;
 
     opened: boolean;
@@ -56,14 +65,27 @@ export class KeybindsComponent implements OnInit {
 
     editingName: boolean = false;
 
+    private _unsubscribeAll: Subject<any> = new Subject<any>();
+
     /**
      * Constructor
      */
-    constructor(private keybindingService: KeybindingService, private _formBuilder: FormBuilder) {
+    constructor(
+        private keybindingService: KeybindingService,
+        private _formBuilder: FormBuilder,
+        private _authService: AuthService,
+        private dialog: MatDialog) {
         this.keybindingSelected = false;
     }
 
     ngOnInit(): void {
+
+        this._authService.check()
+            .pipe(takeUntil(this._unsubscribeAll))
+            .subscribe((authenticated) => {
+                this.isAuthenticated = authenticated;
+            })
+
         this.opened = true;
 
         this.nameForm = this._formBuilder.group({
@@ -74,10 +96,18 @@ export class KeybindsComponent implements OnInit {
 
     onKeybindingSelected(keybinding: any) {
         console.log('onKeybindingSelected', keybinding)
-        this.selectedKeybinding = keybinding; // Handle the emitted keybind from the child component
-        this.selectedKeybindingName = this.selectedKeybinding.name;
-        this.keybindingSelected = true;
-        this.nameForm.get('name')?.setValue(this.selectedKeybindingName);
+        if (keybinding) {
+            this.selectedKeybinding = keybinding; // Handle the emitted keybind from the child component
+            this.selectedKeybindingName = this.selectedKeybinding.name;
+            this.keybindingSelected = true;
+            this.nameForm.get('name')?.setValue(this.selectedKeybindingName);
+        }
+        else {
+            this.selectedKeybinding = null;
+            this.selectedKeybindingName = null;
+            this.keybindingSelected = false;
+            this.nameForm.get('name')?.setValue('');
+        }
     }
 
     onSelectionClassChanged(value: string) {
@@ -86,10 +116,21 @@ export class KeybindsComponent implements OnInit {
     }
 
     deleteKeybinding() {
-        console.log('deleting keybinding', this.selectedKeybinding);
-        this.keybindingService.removeKeybinding(this.selectedKeybinding.id);
-        this.selectedKeybinding = null;
-        this.keybindingSelected = false;
+
+        const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+            data: { text: 'Are you sure you want to delete?' }
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                console.log('deleting keybinding', this.selectedKeybinding);
+                this.keybindingService.removeKeybinding(this.selectedKeybinding.id);
+                this.selectedKeybinding = null;
+                this.keybindingSelected = false;
+            } else {
+                console.log('Selection cancelled');
+            }
+        });
     }
 
     saveKeybinding() {
